+++ client/keys/utils.go
@@ -1,7 +1,6 @@
 package keys
 
 import (
-	"encoding/json"
 	"fmt"
 	"path/filepath"
 
@@ -90,7 +89,7 @@ func printInfo(info keys.Info) {
 		fmt.Printf("NAME:\tADDRESS:\t\t\t\t\t\tPUBKEY:\n")
 		printKeyOutput(ko)
 	case "json":
-		out, err := json.MarshalIndent(ko, "", "\t")
+		out, err := MarshalJSON(ko)
 		if err != nil {
 			panic(err)
 		}
@@ -110,7 +109,7 @@ func printInfos(infos []keys.Info) {
 			printKeyOutput(ko)
 		}
 	case "json":
-		out, err := json.MarshalIndent(kos, "", "\t")
+		out, err := MarshalJSON(kos)
 		if err != nil {
 			panic(err)
 		}