+++ x/stake/genesis.go
@@ -22,21 +22,32 @@ func NewGenesisState(pool Pool, params Params, validators []Validator, bonds []D
 // get raw genesis raw message for testing
 func DefaultGenesisState() GenesisState {
 	return GenesisState{
-		Pool:   initialPool(),
-		Params: defaultParams(),
+		Pool:   InitialPool(),
+		Params: DefaultParams(),
 	}
 }
 
 // InitGenesis - store genesis parameters
 func InitGenesis(ctx sdk.Context, k Keeper, data GenesisState) {
+	store := ctx.KVStore(k.storeKey)
 	k.setPool(ctx, data.Pool)
 	k.setNewParams(ctx, data.Params)
 	for _, validator := range data.Validators {
-		k.updateValidator(ctx, validator)
+
+		// set validator
+		k.setValidator(ctx, validator)
+
+		// manually set indexes for the first time
+		k.setValidatorByPubKeyIndex(ctx, validator)
+		k.setValidatorByPowerIndex(ctx, validator, data.Pool)
+		if validator.Status() == sdk.Bonded {
+			store.Set(GetValidatorsBondedKey(validator.PubKey), validator.Owner)
+		}
 	}
 	for _, bond := range data.Bonds {
 		k.setDelegation(ctx, bond)
 	}
+	k.updateBondedValidatorsFull(ctx, store)
 }
 
 // WriteGenesis - output genesis parameters