+++ x/stake/pool.go
@@ -25,13 +25,13 @@ type Pool struct {
 }
 
 func (p Pool) equal(p2 Pool) bool {
-	bz1 := cdcEmpty.MustMarshalBinary(&p)
-	bz2 := cdcEmpty.MustMarshalBinary(&p2)
+	bz1 := msgCdc.MustMarshalBinary(&p)
+	bz2 := msgCdc.MustMarshalBinary(&p2)
 	return bytes.Equal(bz1, bz2)
 }
 
 // initial pool for testing
-func initialPool() Pool {
+func InitialPool() Pool {
 	return Pool{
 		LooseUnbondedTokens:     0,
 		BondedTokens:            0,