+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## Pending
+
+BREAKING CHANGES
+* [cli] rearranged commands under subcommands
+
+FEATURES
+
+IMPROVEMENTS
+* bank module uses go-wire codec instead of 'encoding/json'
+* auth module uses go-wire codec instead of 'encoding/json'
+
+FIXES
+* [cli] fixed cli-bash tests
+* [ci] added cli-bash tests
+
 ## 0.18.1
 
 BREAKING CHANGES
@@ -56,6 +71,7 @@ FEATURES
 * [stake] Creation of a validator/delegation generics in `/types`
 * [stake] Helper Description of the store in x/stake/store.md
 * [stake] removed use of caches in the stake keeper
+* [stake] Added REST API
 * [Makefile] Added terraform/ansible playbooks to easily create remote testnets on Digital Ocean
 
 BUG FIXES
@@ -87,7 +103,7 @@ FEATURES
 * [gaiacli] Support queries for candidates, delegator-bonds
 * [gaiad] Added `gaiad export` command to export current state to JSON
 * [x/bank] Tx tags with sender/recipient for indexing & later retrieval
-* [x/stake] Tx tags with delegator/candidate for delegation & unbonding, and candidate info for declare candidate / edit candidacy
+* [x/stake] Tx tags with delegator/candidate for delegation & unbonding, and candidate info for declare candidate / edit validator
 
 IMPROVEMENTS
 
@@ -121,7 +137,7 @@ BREAKING CHANGES
 
 FEATURES:
 
-* Gaia stake commands include, DeclareCandidacy, EditCandidacy, Delegate, Unbond
+* Gaia stake commands include, CreateValidator, EditValidator, Delegate, Unbond
 * MountStoreWithDB without providing a custom store works.
 * Repo is now lint compliant / GoMetaLinter with tendermint-lint integrated into CI
 * Better key output, pubkey go-amino hex bytes now output by default